feat(docs): add pure markdown render mode via renderStyle option
