+++ crates/ox_content_napi/src/lib.rs
@@ -27,8 +27,9 @@ use ox_content_docs::{
     DocItemKind, DocTag, DocsDiagnostic, DocsDiagnosticCode, DocsNavItem, DocsOutputOptions,
     EntryPointDocsOptions, EntryPointSpec, ExportGraph, ExportKind, ExportSource,
     ExternalDocsOptions, ExternalPackageSource, ExtractedDocModule, GraphOptions,
-    MarkdownDocsOptions, MarkdownLinkStyle, MarkdownPathStrategy, NormalizedDocEntry,
-    NormalizedMember, NormalizedParamDoc, NormalizedReturnDoc, ParamDoc, PublicExport,
+    MarkdownDocsOptions, MarkdownLinkStyle, MarkdownPathStrategy, MarkdownRenderStyle,
+    NormalizedDocEntry, NormalizedMember, NormalizedParamDoc, NormalizedReturnDoc, ParamDoc,
+    PublicExport,
 };
 use ox_content_parser::{Parser, ParserOptions};
 use ox_content_renderer::HtmlRenderer;
@@ -280,6 +281,8 @@ pub struct JsDocsMarkdownOptions {
     pub base_path: Option<String>,
     #[napi(ts_type = "'flat' | 'typedoc'")]
     pub path_strategy: Option<String>,
+    #[napi(ts_type = "'html' | 'markdown'")]
+    pub render_style: Option<String>,
 }
 
 /// Options for writing generated API documentation files.
@@ -1102,6 +1105,7 @@ pub fn generate_docs_markdown(
             link_style: parse_markdown_link_style(options.link_style.as_deref()),
             base_path: options.base_path,
             path_strategy: parse_markdown_path_strategy(options.path_strategy.as_deref()),
+            render_style: parse_markdown_render_style(options.render_style.as_deref()),
         });
     generate_markdown(&docs.into_iter().map(convert_markdown_module).collect::<Vec<_>>(), &options)
         .into_iter()
@@ -1119,6 +1123,13 @@ fn parse_markdown_path_strategy(path_strategy: Option<&str>) -> MarkdownPathStra
     match path_strategy {
         Some("typedoc") => MarkdownPathStrategy::TypeDoc,
         _ => MarkdownPathStrategy::Flat,
+    }
+}
+
+fn parse_markdown_render_style(render_style: Option<&str>) -> MarkdownRenderStyle {
+    match render_style {
+        Some("markdown") => MarkdownRenderStyle::Markdown,
+        _ => MarkdownRenderStyle::Html,
     }
 }
 
@@ -3089,12 +3100,52 @@ mod tests {
                 link_style: Some("clean".to_string()),
                 base_path: Some("/api-ox".to_string()),
                 path_strategy: None,
+                render_style: None,
             }),
         );
         let index = markdown.get("index.md").unwrap();
 
         assert!(index.contains("href=\"/api-ox/context\""));
         assert!(index.contains("href=\"/api-ox/context#commandcontext\""));
+    }
+
+    #[test]
+    fn generate_docs_markdown_render_style_markdown_omits_html() {
+        let docs = vec![JsDocsMarkdownModule {
+            file: "/repo/src/context.ts".to_string(),
+            entries: vec![JsDocsMarkdownEntry {
+                name: "CommandContext".to_string(),
+                kind: "interface".to_string(),
+                description: "Runtime context.".to_string(),
+                params: None,
+                returns: None,
+                examples: None,
+                tags: None,
+                private: false,
+                file: "/repo/src/context.ts".to_string(),
+                line: 1,
+                end_line: 1,
+                signature: Some("export interface CommandContext".to_string()),
+                members: None,
+            }],
+        }];
+        let markdown = generate_docs_markdown(
+            docs,
+            Some(JsDocsMarkdownOptions {
+                group_by: Some("file".to_string()),
+                github_url: None,
+                link_style: None,
+                base_path: None,
+                path_strategy: None,
+                render_style: Some("markdown".to_string()),
+            }),
+        );
+        let page = markdown.get("context.md").unwrap();
+
+        assert!(!page.contains("<details"));
+        assert!(!page.contains("class=\"ox-api"));
+        assert!(page.contains("### CommandContext"));
+        assert!(page.contains("```ts"));
     }
 
     #[test]
@@ -3228,6 +3279,7 @@ mod tests {
                 link_style: Some("clean".to_string()),
                 base_path: Some("/api".to_string()),
                 path_strategy: Some("typedoc".to_string()),
+                render_style: None,
             }),
         );
         let cli_page = markdown.get("default/functions/cli.md").unwrap();
@@ -3345,6 +3397,7 @@ mod tests {
                 link_style: Some("clean".to_string()),
                 base_path: Some("/api".to_string()),
                 path_strategy: Some("typedoc".to_string()),
+                render_style: None,
             }),
         );
